Use server component for better SEO
